# contracts/openapi.yaml
openapi: 3.0.0

info:
  title: Zerodha Clone API
  version: 1.0.0
  description: API contracts for Zerodha Clone application

servers:
  - url: http://backend:3002
  - url: http://localhost:3002

paths:

  /auth/signup:
    post:
      operationId: signupUser
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  pattern: '^[A-Za-z\d]{6,}$'
                  example: priya123
                email:
                  type: string
                  example: priya@gmail.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: User registered and logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      operationId: loginUser
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: priya123
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      operationId: logoutUser
      summary: Logout current user
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out

  /auth/me:
    get:
      operationId: getMe
      summary: Get currently logged-in user
      security:
        - testAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: string
                    nullable: true
                    example: null

  /auth/check-username:
    get:
      operationId: checkUsername
      summary: Check if a username is available
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
          example: priya123
      responses:
        '200':
          description: Availability result
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    example: true
        '400':
          description: Missing username query parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                    example: false

  /allHoldings:
    get:
      operationId: getAllHoldings
      summary: Get all stock holdings
      responses:
        '200':
          description: List of holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holding'

  /allPositions:
    get:
      operationId: getAllPositions
      summary: Get all trading positions
      responses:
        '200':
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'

  /newOrder:
    post:
      operationId: placeNewOrder
      summary: Place a new stock order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - qty
                - price
                - mode
              properties:
                name:
                  type: string
                  example: INFY
                qty:
                  type: integer
                  minimum: 1
                  example: 5
                price:
                  type: number
                  example: 1555.45
                mode:
                  type: string
                  enum:
                    - BUY
                    - SELL
                  example: BUY
      responses:
        '200':
          description: Order placed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order saved!
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /create-order:
    post:
      operationId: createOrder
      summary: Create a Razorpay payment order for adding funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  example: 5000
      responses:
        '200':
          description: Razorpay order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RazorpayOrder'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Order creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /verify-payment:
    post:
      operationId: verifyPayment
      summary: Verify Razorpay payment and credit funds to user balance
      security:
        - testAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - razorpay_order_id
                - razorpay_payment_id
                - razorpay_signature
                - amount
              properties:
                razorpay_order_id:
                  type: string
                  example: order_abc123
                razorpay_payment_id:
                  type: string
                  example: pay_xyz789
                razorpay_signature:
                  type: string
                  example: abc123sig
                amount:
                  type: number
                  example: 5000
      responses:
        '200':
          description: Payment verified and balance updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  balance:
                    type: number
                    example: 55000
        '400':
          description: Signature verification failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failure
        '401':
          description: Login required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    testAuth:
      type: apiKey
      in: header
      name: X-Test-Username
      description: Test-only header that identifies the acting user; backend resolves it to a session in non-production environments.

  schemas:

    User:
      type: object
      additionalProperties: true
      properties:
        _id:
          type: string
          example: 64abc123def456
        username:
          type: string
          example: priya123
        email:
          type: string
          example: priya@gmail.com
        balance:
          type: number
          example: 50000

    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'

    Holding:
      type: object
      properties:
        _id:
          type: string
          example: 64abc123def456
        name:
          type: string
          example: TCS
        qty:
          type: number
          example: 1
        avg:
          type: number
          example: 3041.70
        price:
          type: number
          example: 3194.80
        net:
          type: string
          example: "+5.03%"
        day:
          type: string
          example: "-0.25%"
        sector:
          type: string
          example: "Technology"

    Position:
      type: object
      required:
        - name
        - qty
        - price
      properties:
        _id:
          type: string
          example: 64abc123def456
        product:
          type: string
          example: CNC
        name:
          type: string
          example: EVEREADY
        qty:
          type: number
          example: 2
        avg:
          type: number
          example: 316.27
        price:
          type: number
          example: 312.35
        net:
          type: string
          example: "+0.58%"
        day:
          type: string
          example: "-1.24%"
        isLoss:
          type: boolean
          example: true

    RazorpayOrder:
      type: object
      required:
        - id
        - amount
      properties:
        id:
          type: string
          example: order_abc123xyz
        entity:
          type: string
          example: order
        amount:
          type: number
          example: 500000
        amount_paid:
          type: number
          example: 0
        amount_due:
          type: number
          example: 500000
        currency:
          type: string
          example: INR
        receipt:
          type: string
          example: funds_1716000000000
        status:
          type: string
          example: created
        attempts:
          type: number
          example: 0
        notes:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: number
          example: 1716000000
        offer_id:
          type: string
          nullable: true
          example: null

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong